(refactor) rename and create files for modularity
