+++ client/src/actions/changecashbackpercent.js
@@ -15,14 +15,13 @@ export const decrementCashbackPercent = (ccid, catid) => {
     ccid: ccid,
     catid: catid
   };
-}
+};
 
 export const changeCashbackPercentError = () => {
   return {
     type: 'CHANGE_CASHBACK_PERCENT_ERROR'
   };
-}
-
+};
 
 export const changeCashbackPercent = (catid, percent, action, ccid) => {
   return (dispatch) => {
@@ -49,48 +48,6 @@ export const changeCashbackPercent = (catid, percent, action, ccid) => {
     .catch((err) => {
       console.log('error in get', err);
       dispatch(changeCashbackPercentError(err));
-    });
-  };
-};
-
-export const createCashbackCategoryError = () => {
-  return {
-    type: 'CREATE_CASHBACK_CATEGORY_ERROR'
-  };
-}
-
-export const createCashbackCategory = (ccid, name, percent, catid) => {
-  return {
-    type: 'CREATE_CASHBACK_CATEGORY',
-    ccid: ccid, 
-    name: name,
-    percent: percent,
-    catid: catid
-  };
-}
-
-export const createCashbackCategoryKickoff = (ccid, name, percent) => {
-  return (dispatch) => {
-    fetch('/creditcards/createcashbackcategory', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      credentials: 'same-origin',
-      body: JSON.stringify({
-        ccid: ccid, 
-        name: name, 
-        percent: percent
-      })
-    })
-    .then(response => {
-      console.log('response', response);
-      var catid = response.insertId;
-      dispatch(createCashbackCategory(ccid, name, percent, catid));
-    })
-    .catch((err) => {
-      console.log('error in get', err);
-      dispatch(createCashbackCategoryError(err));
     });
   };
 };
